First Steps of Refactoring ClientService.cs
